requirements.txt
flask
elasticsearch
numpy
tqdm
pillow
tensorflow==2.4.1
